backend/requirements.txt
flask
flask-cors
SQLAlchemy
pytest
pytest-flask
reportlab
pypdf
Werkzeug
flasgger
python-dotenv
